+++ SQL_Server/DTOs/BusinessAssociatePhoneDTO.cs
@@ -2,7 +2,7 @@ namespace SQL_Server.DTOs
 {
     public class BusinessAssociatePhoneDTO
     {
-        public required int BusinessAssociate_Legal_Id { get; set; } // PK, FK
-        public required int Phone { get; set; } // PK
+        public required long BusinessAssociate_Legal_Id { get; set; } // PK, FK
+        public required long Phone { get; set; } // PK
     }
 }
