+++ SQL_Server/DTOs/BusinessManagerPhoneDTO.cs
@@ -3,6 +3,6 @@ namespace SQL_Server.DTOs
     public class BusinessManagerPhoneDTO
     {
         public required string BusinessManager_Email { get; set; } // PK, FK
-        public required int Phone { get; set; } // PK
+        public required long Phone { get; set; } // PK
     }
 }
