+++ SQL_Server/DTOs/CartDTO.cs
@@ -2,9 +2,9 @@ namespace SQL_Server.DTOs
 {
     public class CartDTO
     {
-        public int Code { get; set; } // PK
-        public int? BusinessAssociate_Legal_Id { get; set; }
-        public int? TotalProductsPrice { get; set; }
-        public required int Client_Id { get; set; } // FK
+        public long Code { get; set; } // PK
+        public long? BusinessAssociate_Legal_Id { get; set; }
+        public long? TotalProductsPrice { get; set; }
+        public required long Client_Id { get; set; } // FK
     }
 }
